Move schema to `config` crate
